+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/RepositoryImportResource.java
@@ -32,8 +32,6 @@ import com.google.common.base.Strings;
 import com.google.common.io.ByteSource;
 import com.google.common.io.Files;
 import com.google.inject.Inject;
-import de.otto.edison.hal.Embedded;
-import de.otto.edison.hal.Links;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -47,10 +45,12 @@ import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
 import org.jboss.resteasy.plugins.providers.multipart.MultipartInputImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import sonia.scm.ContextEntry;
 import sonia.scm.HandlerEventType;
 import sonia.scm.Type;
 import sonia.scm.event.ScmEventBus;
 import sonia.scm.importexport.FullScmRepositoryImporter;
+import sonia.scm.importexport.RepositoryImportExportEncryption;
 import sonia.scm.repository.InternalRepositoryException;
 import sonia.scm.repository.Repository;
 import sonia.scm.repository.RepositoryImportEvent;
@@ -58,6 +58,7 @@ import sonia.scm.repository.RepositoryManager;
 import sonia.scm.repository.RepositoryPermission;
 import sonia.scm.repository.RepositoryPermissions;
 import sonia.scm.repository.api.Command;
+import sonia.scm.repository.api.ImportFailedException;
 import sonia.scm.repository.api.PullCommandBuilder;
 import sonia.scm.repository.api.RepositoryService;
 import sonia.scm.repository.api.RepositoryServiceFactory;
@@ -93,7 +94,6 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.singletonList;
 import static sonia.scm.api.v2.resources.RepositoryTypeSupportChecker.checkSupport;
 import static sonia.scm.api.v2.resources.RepositoryTypeSupportChecker.type;
-import static sonia.scm.importexport.RepositoryImportExportEncryption.decrypt;
 
 public class RepositoryImportResource {
 
@@ -105,6 +105,7 @@ public class RepositoryImportResource {
   private final ResourceLinks resourceLinks;
   private final ScmEventBus eventBus;
   private final FullScmRepositoryImporter fullScmRepositoryImporter;
+  private final RepositoryImportExportEncryption repositoryImportExportEncryption;
 
   @Inject
   public RepositoryImportResource(RepositoryManager manager,
@@ -112,13 +113,15 @@ public class RepositoryImportResource {
                                   RepositoryServiceFactory serviceFactory,
                                   ResourceLinks resourceLinks,
                                   ScmEventBus eventBus,
-                                  FullScmRepositoryImporter fullScmRepositoryImporter) {
+                                  FullScmRepositoryImporter fullScmRepositoryImporter,
+                                  RepositoryImportExportEncryption repositoryImportExportEncryption) {
     this.manager = manager;
     this.mapper = mapper;
     this.serviceFactory = serviceFactory;
     this.resourceLinks = resourceLinks;
     this.eventBus = eventBus;
     this.fullScmRepositoryImporter = fullScmRepositoryImporter;
+    this.repositoryImportExportEncryption = repositoryImportExportEncryption;
   }
 
   /**
@@ -353,10 +356,11 @@ public class RepositoryImportResource {
   }
 
   private InputStream decryptInputStream(InputStream inputStream, String password) {
-    if (!Strings.isNullOrEmpty(password)) {
-      inputStream = decrypt(inputStream, password);
+    try {
+      return repositoryImportExportEncryption.decrypt(inputStream, password);
+    } catch (IOException e) {
+      throw new ImportFailedException(ContextEntry.ContextBuilder.noContext(), "import failed", e);
     }
-    return inputStream;
   }
 
   @VisibleForTesting
@@ -427,15 +431,10 @@ public class RepositoryImportResource {
   @NoArgsConstructor
   @SuppressWarnings("java:S2160")
   public static class RepositoryImportFromUrlDto extends RepositoryDto implements ImportRepositoryFromUrlDto {
-
     @NotEmpty
     private String importUrl;
     private String username;
     private String password;
-
-    RepositoryImportFromUrlDto(Links links, Embedded embedded) {
-      super(links, embedded);
-    }
   }
 
   @Getter
@@ -444,10 +443,6 @@ public class RepositoryImportResource {
   @SuppressWarnings("java:S2160")
   public static class RepositoryImportFromFileDto extends RepositoryDto implements ImportRepositoryFromFileDto {
     private String password;
-
-    RepositoryImportFromFileDto(Links links, Embedded embedded) {
-      super(links, embedded);
-    }
   }
 
   interface ImportRepositoryDto {