+++ scm-webapp/src/main/java/sonia/scm/importexport/FullScmRepositoryImporter.java
@@ -33,8 +33,10 @@ import sonia.scm.repository.Repository;
 import sonia.scm.repository.RepositoryManager;
 import sonia.scm.repository.RepositoryPermission;
 import sonia.scm.repository.api.ImportFailedException;
+import sonia.scm.repository.api.IncompatibleEnvironmentForImportException;
 import sonia.scm.repository.api.RepositoryService;
 import sonia.scm.repository.api.RepositoryServiceFactory;
+import sonia.scm.update.UpdateEngine;
 
 import javax.inject.Inject;
 import javax.xml.bind.JAXB;
@@ -49,26 +51,32 @@ import static sonia.scm.ContextEntry.ContextBuilder.noContext;
 import static sonia.scm.importexport.FullScmRepositoryExporter.METADATA_FILE_NAME;
 import static sonia.scm.importexport.FullScmRepositoryExporter.SCM_ENVIRONMENT_FILE_NAME;
 import static sonia.scm.importexport.FullScmRepositoryExporter.STORE_DATA_FILE_NAME;
-import static sonia.scm.importexport.RepositoryImportExportEncryption.decrypt;
 
 public class FullScmRepositoryImporter {
 
+  @SuppressWarnings("java:S115") // we like this name here
   private static final int _1_MB = 1000000;
 
   private final RepositoryServiceFactory serviceFactory;
   private final RepositoryManager repositoryManager;
   private final ScmEnvironmentCompatibilityChecker compatibilityChecker;
   private final TarArchiveRepositoryStoreImporter storeImporter;
+  private final UpdateEngine updateEngine;
+  private final RepositoryImportExportEncryption repositoryImportExportEncryption;
 
   @Inject
   public FullScmRepositoryImporter(RepositoryServiceFactory serviceFactory,
                                    RepositoryManager repositoryManager,
                                    ScmEnvironmentCompatibilityChecker compatibilityChecker,
-                                   TarArchiveRepositoryStoreImporter storeImporter) {
+                                   TarArchiveRepositoryStoreImporter storeImporter,
+                                   UpdateEngine updateEngine,
+                                   RepositoryImportExportEncryption repositoryImportExportEncryption) {
     this.serviceFactory = serviceFactory;
     this.repositoryManager = repositoryManager;
     this.compatibilityChecker = compatibilityChecker;
     this.storeImporter = storeImporter;
+    this.updateEngine = updateEngine;
+    this.repositoryImportExportEncryption = repositoryImportExportEncryption;
   }
 
   public Repository importFromStream(Repository repository, InputStream inputStream, String password) {
@@ -76,7 +84,7 @@ public class FullScmRepositoryImporter {
       if (inputStream.available() > 0) {
         try (
           BufferedInputStream bif = new BufferedInputStream(inputStream);
-          InputStream cif = decrypt(bif, password);
+          InputStream cif = repositoryImportExportEncryption.decrypt(bif, password);
           GzipCompressorInputStream gcis = new GzipCompressorInputStream(cif);
           TarArchiveInputStream tais = new TarArchiveInputStream(gcis)
         ) {
@@ -116,6 +124,7 @@ public class FullScmRepositoryImporter {
       // Inside the repository tar archive stream is another tar archive.
       // The nested tar archive is wrapped in another TarArchiveInputStream inside the storeImporter
       storeImporter.importFromTarArchive(repository, tais);
+      updateEngine.update(repository.getId());
     } else {
       throw new ImportFailedException(
         ContextEntry.ContextBuilder.entity(repository).build(),
@@ -151,10 +160,7 @@ public class FullScmRepositoryImporter {
     if (environmentEntry.getName().equals(SCM_ENVIRONMENT_FILE_NAME) && !environmentEntry.isDirectory() && environmentEntry.getSize() < _1_MB) {
       boolean validEnvironment = compatibilityChecker.check(JAXB.unmarshal(new NoneClosingInputStream(tais), ScmEnvironment.class));
       if (!validEnvironment) {
-        throw new ImportFailedException(
-          noContext(),
-          "Incompatible SCM-Manager environment. Could not import file."
-        );
+        throw new IncompatibleEnvironmentForImportException();
       }
     } else {
       throw new ImportFailedException(